ui + approval handling
